// client/src/autobindings/types.ts
export const ActiveMonsterAttack = __t.object("ActiveMonsterAttack", {
  activeMonsterAttackId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
  get position() {
    return DbVector2;
  },
  get direction() {
    return DbVector2;
  },
  get monsterAttackType() {
    return MonsterAttackType;
  },
  piercing: __t.bool(),
  damage: __t.u32(),
  radius: __t.f32(),
  speed: __t.f32(),
  parameterU: __t.u32(),
  parameterF: __t.f32(),
  ticksElapsed: __t.u32(),
  fromShinyMonster: __t.bool(),
});
export type ActiveMonsterAttack = __Infer<typeof ActiveMonsterAttack>;
// ...
export const AgnaMagicCircle = __t.object("AgnaMagicCircle", {
  circleId: __t.u64(),
  bossMonsterId: __t.u32(),
  targetPlayerId: __t.u32(),
  circleIndex: __t.u32(),
  initialRotation: __t.f32(),
  ticksElapsed: __t.u32(),
  get position() {
    return DbVector2;
  },
});
export type AgnaMagicCircle = __Infer<typeof AgnaMagicCircle>;
// ...
export const AiState = __t.enum("AiState", {
  Default: __t.unit(),
  BossEnderIdle: __t.unit(),
  BossEnderChase: __t.unit(),
  BossEnderDance: __t.unit(),
  BossEnderVanish: __t.unit(),
  BossEnderLurk: __t.unit(),
  BossEnderTeleport: __t.unit(),
  BossEnderTransform: __t.unit(),
  Stationary: __t.unit(),
  BossAgnaIdle: __t.unit(),
  BossAgnaFlamethrower: __t.unit(),
  BossAgnaMagicCircle: __t.unit(),
  BossAgnaRitualMatch: __t.unit(),
  BossAgnaRitualWick: __t.unit(),
  BossAgnaRitualFailed: __t.unit(),
  BossAgnaRitualComplete: __t.unit(),
  BossSimonIdle: __t.unit(),
  BossSimonToxicSpray: __t.unit(),
  BossSimonZombieWave: __t.unit(),
  BossSimonTransform: __t.unit(),
  BossSimonChemicalBoltPattern: __t.unit(),
  BossSimonToxicZonePattern: __t.unit(),
  BossSimonPhase2Transform: __t.unit(),
});
export type AiState = __Infer<typeof AiState>;
// ...
export const DbVector2 = __t.object("DbVector2", {
  x: __t.f32(),
  y: __t.f32(),
});
export type DbVector2 = __Infer<typeof DbVector2>;
// ...
export const MonsterAttackType = __t.enum("MonsterAttackType", {
  ImpBolt: __t.unit(),
  EnderBolt: __t.unit(),
  EnderScytheSpawn: __t.unit(),
  EnderScythe: __t.unit(),
  ChaosBall: __t.unit(),
  VoidZone: __t.unit(),
  AgnaFlamethrowerJet: __t.unit(),
  AgnaOrbSpawn: __t.unit(),
  AgnaFireOrb: __t.unit(),
  AgnaCandleBolt: __t.unit(),
  AgnaPhase2FlameJet: __t.unit(),
  AgnaGroundFlame: __t.unit(),
  SimonChemicalBolt: __t.unit(),
  SimonToxicZone: __t.unit(),
  SimonToxicSpray: __t.unit(),
});
export type MonsterAttackType = __Infer<typeof MonsterAttackType>;
// ...
export const MonsterType = __t.enum("MonsterType", {
  Rat: __t.unit(),
  Slime: __t.unit(),
  Bat: __t.unit(),
  Orc: __t.unit(),
  Imp: __t.unit(),
  Zombie: __t.unit(),
  VoidChest: __t.unit(),
  EnderClaw: __t.unit(),
  BossEnderPhase1: __t.unit(),
  BossEnderPhase2: __t.unit(),
  BossAgnaPhase1: __t.unit(),
  BossAgnaPhase2: __t.unit(),
  BossSimonPhase1: __t.unit(),
  BossSimonPhase2: __t.unit(),
  AgnaCandle: __t.unit(),
  Crate: __t.unit(),
  Tree: __t.unit(),
  Statue: __t.unit(),
});
export type MonsterType = __Infer<typeof MonsterType>;

// The tagged union or sum type for the algebraic type `MonsterVariant`.
export const MonsterVariant = __t.enum("MonsterVariant", {
  Default: __t.unit(),
  Shiny: __t.unit(),
  Cursed: __t.unit(),
});
export type MonsterVariant = __Infer<typeof MonsterVariant>;

export const Monsters = __t.object("Monsters", {
  monsterId: __t.u32(),
  get bestiaryId() {
    return MonsterType;
  },
  get variant() {
    return MonsterVariant;
  },
  hp: __t.u32(),
  maxHp: __t.u32(),
  atk: __t.f32(),
  speed: __t.f32(),
  targetPlayerId: __t.u32(),
  get aiState() {
    return AiState;
  },
  radius: __t.f32(),
  get spawnPosition() {
    return DbVector2;
  },
});
export type Monsters = __Infer<typeof Monsters>;

export const Player = __t.object("Player", {
  playerId: __t.u32(),
  name: __t.string(),
  spawnGracePeriodRemaining: __t.u32(),
  get playerClass() {
    return PlayerClass;
  },
  level: __t.u32(),
  exp: __t.u32(),
  expForNextLevel: __t.u32(),
  maxHp: __t.f32(),
  hp: __t.f32(),
  hpRegen: __t.u32(),
  speed: __t.f32(),
  armor: __t.u32(),
  unspentUpgrades: __t.u32(),
  rerolls: __t.u32(),
  shieldCount: __t.u32(),
  pvp: __t.bool(),
  get waypoint() {
    return DbVector2;
  },
  hasWaypoint: __t.bool(),
  get position() {
    return DbVector2;
  },
  radius: __t.f32(),
  isBot: __t.bool(),
});
export type Player = __Infer<typeof Player>;

// The tagged union or sum type for the algebraic type `PlayerClass`.
export const PlayerClass = __t.enum("PlayerClass", {
  Fighter: __t.unit(),
  Rogue: __t.unit(),
  Mage: __t.unit(),
  Paladin: __t.unit(),
  Valkyrie: __t.unit(),
  Priest: __t.unit(),
  Football: __t.unit(),
  Gambler: __t.unit(),
  Athlete: __t.unit(),
  Gourmand: __t.unit(),
  Volleyball: __t.unit(),
  Stoner: __t.unit(),
});
export type PlayerClass = __Infer<typeof PlayerClass>;